handled no instance case

+++ src/neurocaas_contrib/cli_commands.py
@@ -61,23 +61,26 @@ def delete_ami_from_cli(develop_dict,force = False):
     from .remote import NeuroCAASAMI
     analysis = develop_dict["config"]["PipelineName"]
     instance = develop_dict["instance_id"]
-    click.confirm("Detected an existing development session with instance {} for analysis {}. Delete session?".format(analysis,instance),abort = True)
-    ## delete instance.
-    ami = NeuroCAASAMI.from_dict(develop_dict)
-    message = ami.terminate_devinstance(force)
-    if message == "No state change.":
-        return False
-    else:
-        try:
-            with open(configpath,"r") as f:
-                config = json.load(f)
-            config["develop_dict"] = None    
-        except FileNotFoundError:    
-            click.echo("NeuroCAAS Contrib config file not found. Exiting.")
-            raise
-        with open(configpath,"w") as f:
-            json.dump(config,f,indent=4)
-        return True    
+    if instance is not None: 
+        click.confirm("Detected an existing development session with instance {} for analysis {}. Delete session?".format(analysis,instance),abort = True)
+        ## delete instance.
+        ami = NeuroCAASAMI.from_dict(develop_dict)
+        message = ami.terminate_devinstance(force)
+        if message == "No state change.":
+            return False
+        else:
+            try:
+                with open(configpath,"r") as f:
+                    config = json.load(f)
+                config["develop_dict"] = None    
+            except FileNotFoundError:    
+                click.echo("NeuroCAAS Contrib config file not found. Exiting.")
+                raise
+            with open(configpath,"w") as f:
+                json.dump(config,f,indent=4)
+            return True    
+    else:     
+        click.echo("No development instance detected. Resetting session.")
 
 def create_ctx(ctx,location,analysis_name,develop_dict):
     """helper function to attempt to create as much of the context object as is available.
